Extract initialization module - GOAL ACHIEVED: 991 lines (84% reduction from 6,274 lines)
